requirements.txt
openai
fastapi
uvicorn[standard]
gunicorn
python-dotenv
pydantic
httpx
streamlit
faiss-cpu
numpy
tiktoken
reportlab
python-docx
